+++ tools/cli/Cargo.toml
@@ -14,6 +14,7 @@ repository = { workspace=true }
 
 [features]
 default = ["unix"]
+idv-tui = ["dep:cursive"]
 unix = []
 
 [lib]
@@ -50,9 +51,12 @@ tokio = { workspace = true, features = ["rt", "macros"] }
 url = { workspace = true, features = ["serde"] }
 uuid = { workspace=true }
 zxcvbn = { workspace=true }
+lazy_static.workspace = true
+regex.workspace = true
 
 [dependencies.cursive]
 version = "0.20.0"
+optional = true
 default-features = false
 features = ["crossterm-backend"]
